+++ dispatch_tables_arm64_test.go
@@ -20,6 +20,7 @@ var allFlatTables = map[string][]any{
 	"Bytes.Adler32":                  {tblBytesAdler32[0], tblBytesAdler32[1], tblBytesAdler32[2]},
 	"Bytes.CRC32C":                   {tblBytesCRC32C[0], tblBytesCRC32C[1], tblBytesCRC32C[2]},
 	"Bytes.DtoaF64":                  {tblBytesDtoaF64[0], tblBytesDtoaF64[1], tblBytesDtoaF64[2]},
+	"Bytes.LZ4BlockDecode":           {tblBytesLZ4BlockDecode[0], tblBytesLZ4BlockDecode[1], tblBytesLZ4BlockDecode[2]},
 	"Convert.QMatMulI8":              {tblConvertQMatMulI8[0], tblConvertQMatMulI8[1], tblConvertQMatMulI8[2]},
 	"Convert.RequantizeI8":           {tblConvertRequantizeI8[0], tblConvertRequantizeI8[1], tblConvertRequantizeI8[2]},
 	"Mask.All":                       {tblMaskAll[0], tblMaskAll[1], tblMaskAll[2]},