+++ dispatch_tables_fallback.go
@@ -22,6 +22,7 @@ var tblBytesRunStartsU8 = [...]func([]bool, []byte){ref.RunStartsU8}
 var tblBytesAdler32 = [...]func([]byte, uint32) uint32{ref.Adler32}
 var tblBytesCRC32C = [...]func([]byte, uint32) uint32{ref.CRC32C}
 var tblBytesDtoaF64 = [...]func([]byte, float64) int{ref.DtoaF64}
+var tblBytesLZ4BlockDecode = [...]func([]byte, []byte) int{ref.LZ4BlockDecode}
 var tblConvertQMatMulI8 = [...]func([]int32, []int8, []int8, int, int, int){ref.QMatMulI8}
 var tblConvertRequantizeI8 = [...]func([]int8, []int32, float32, int32){ref.RequantizeI8}
 var tblMaskAll = [...]func([]bool) bool{ref.MaskAll}