+++ dispatch_tables_riscv64_test.go
@@ -20,6 +20,7 @@ var allFlatTables = map[string][]any{
 	"Bytes.Adler32":                  {tblBytesAdler32[0], tblBytesAdler32[1]},
 	"Bytes.CRC32C":                   {tblBytesCRC32C[0], tblBytesCRC32C[1]},
 	"Bytes.DtoaF64":                  {tblBytesDtoaF64[0], tblBytesDtoaF64[1]},
+	"Bytes.LZ4BlockDecode":           {tblBytesLZ4BlockDecode[0], tblBytesLZ4BlockDecode[1]},
 	"Convert.QMatMulI8":              {tblConvertQMatMulI8[0], tblConvertQMatMulI8[1]},
 	"Convert.RequantizeI8":           {tblConvertRequantizeI8[0], tblConvertRequantizeI8[1]},
 	"Mask.All":                       {tblMaskAll[0], tblMaskAll[1]},